[US164Task178] Modified after PR review 3
