+++ lead-time/service/simle_service.py
@@ -1,0 +1,20 @@
+class SimpleCache:
+    def __init__(self):
+        self.cache = {}
+
+    def set(self, key, value):
+        self.cache[key] = value
+
+    def get(self, key):
+        return self.cache.get(key)
+
+    def delete(self, key):
+        if key in self.cache:
+            del self.cache[key]
+
+    def clear(self):
+        self.cache.clear()
+
+
+# Create an instance of SimpleCache that can be imported in other files
+cache = SimpleCache()
